Bumped autochrome_junk_drawer to manifest v3
It now uses `declarativeNetRequestWithHostAccess` which 'statically'
modifies requests with rules (useragent_tag_$COLOR.json) depending on
the theme currently loaded.
It also needs the `management` permission in order to determine which
profile/theme is currently being used.
As manifest v2 will not work during 2023, it seems to me as a good time
to upgrade.